requirements.txt
imgaug==0.4.0
kornia==0.5.8
matplotlib==3.1.0
numpy
opencv_contrib_python
Pillow
PyYAML==6.0
scipy
torch
torchvision
torchaudio
tqdm==4.32.1
